requirements.txt
# Core ML Libraries
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.0

# Experiment Tracking
mlflow==2.8.0

# Data Version Control
dvc==3.30.1
dvc-gdrive==2.20.0

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Monitoring & Logging
prometheus-client==0.18.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
matplotlib==3.8.0
seaborn==0.13.0
kaggle==1.6.17

# Container Health
requests==2.31.0
pathspec==0.11.2
